feat(ui): replace mobile settings modal with page
Replaces the mobile settings modal with a new full-page view that slides in from the right, matching the chat page transition.

This change improves the user experience on mobile devices by providing a more native-feeling settings interface.

- Creates a new `MobileSettingsUI.js` component for the page-based view.
- Moves all settings content and logic from the old modal to the new UI component.
- Updates the main UI router in `ui.js` to animate and display the new settings page.
- Deletes the old `MobileSettingsModal.js` and cleans up all related imports and references.
- Adjusts styling of the new settings page, including header, fonts, and icons, based on user feedback.
