+++ frontend/index.html
@@ -28,6 +28,9 @@
 
       <!-- Main Chat Area -->
       <main id="main-chat" class="w-full h-full absolute top-0 left-0 flex flex-col bg-gray-950"></main>
+
+      <!-- Settings Page (for Mobile) -->
+      <div id="settings-page-container" class="w-full h-full absolute top-0 left-0 flex-col bg-gray-950"></div>
     </div>
 
     <!-- Sidebar Backdrop for Mobile -->
