+++ frontend/static/index.css
@@ -347,7 +347,8 @@ textarea::-webkit-scrollbar,
 }
 
 #character-list-page-container,
-#main-chat {
+#main-chat,
+#settings-page-container {
   backface-visibility: hidden;
   transition: transform 0.6s cubic-bezier(0.4, 0, 0.2, 1), opacity 0.6s cubic-bezier(0.4, 0, 0.2, 1);
 }
@@ -365,6 +366,12 @@ textarea::-webkit-scrollbar,
   opacity: 0.8;
 }
 
+#settings-page-container {
+    transform: translateX(100%);
+    z-index: 30; /* Settings on top of chat */
+    opacity: 0.8;
+}
+
 /* "Show Chat" State */
 #page-transition-container.show-chat #character-list-page-container {
   transform: translateX(-40%) scale(0.9);
@@ -372,6 +379,17 @@ textarea::-webkit-scrollbar,
 }
 
 #page-transition-container.show-chat #main-chat {
+  transform: translateX(0);
+  opacity: 1;
+}
+
+/* "Show Settings" State */
+#page-transition-container.show-settings #character-list-page-container {
+  transform: translateX(-40%) scale(0.9);
+  opacity: 0.5;
+}
+
+#page-transition-container.show-settings #settings-page-container {
   transform: translateX(0);
   opacity: 1;
 }
